Ensure JNA always encodes strings as UTF-8.
JNA defaults to encoding strings in the system default encoding,
which is usually UTF-8 but not always. Rust *always* expects strings
encoded in UTF-8, so this commit adds logic to enforce that encoding
when we detect that JNA won't do it for us for free.

I was able to test this locally by setting `LANG=C` in my environment
and then running the Places Kotlin tests. Prior to this commit they
fail on some Unicode test data; with this commit they succeed.

Luckily for us in practice, Android devices *always* use UTF-8 as
the default encoding, so this issue won't have caused any problems
for production users. Still, we shouldn't depend on that in the general
case.